Project reworked for RSocket support
